requirements.txt
numpy>=1.24.0
pandas>=2.0.0
gymnasium>=0.29.0
torch>=2.0.0
matplotlib>=3.7.0
tqdm>=4.65.0
python-dotenv>=1.0.0
kaggle>=1.5.16
